fix bugs and fix description

+++ src/drug_regimen/service.py
@@ -1,3 +1,4 @@
+import logging
 from calendar import monthrange
 from datetime import datetime, timedelta
 
@@ -11,26 +12,35 @@ from src.drug_regimen.state_machine import ManagerState, RegimenState
 from src.drug_regimen.utils import LOCAL_RU_MONTH, get_user_date
 from src.user.service import UserService
 from src.utils.buttons import InlineButtonsGenerator as ibg
+from src.utils.validate_user_message import validate_input
 
 TEMP_DATA_FOR_MANAGER = "temp_data_with_date_for_user"
 
 
 class ManagerService:
+    def __init__(self):
+        self._validate_input = validate_input
+
     async def survey_name(self, message: types.Message, state: FSMContext):
         await state.clear()
         await state.set_state(ManagerState.name)
-        await message.answer(
-            "Придумайте название курсу приема лекарст.\nНазвание может быть произвольным. Главное чтобы вам было понятно.\n"
-            "Кстати это может быть что угодно. Даже курс отжиманий! "
-            "Но, расчитано всё таки со спецификой приема лекарст.\nТак же вкурсе может быть до "
-            "пяти отслеживаний препаратов одновременно.",
+        save_message_data = await message.answer(
+            "Придумайте название курсу.\nНазвание может быть произвольным. Главное чтобы вам было понятно.\n"
+            "Кстати это может быть что угодно. Даже курс отжиманий!\n"
+            "Но, расчитано всё таки со спецификой приема лекарст/витаминов.\nТак же вкурсе может быть до "
+            "пяти отслеживаний препаратов одновременно, а это 4 приема/напоминаний в день на один курс(20 напоминаний день).",
         )
+        await state.update_data(previous_start_message_id=save_message_data.message_id)
 
     async def survey_date(self, message: types.Message, state: FSMContext):
-        await state.set_state(ManagerState.start_date)
-        await state.update_data(plane_name=message.text)
+        if not self._validate_input(message.text):
+            await message.reply("Ваш ввод содержит недопустимые символы. Пожалуйста, попробуйте снова.")
+            return
 
         await message.delete()
+        previous_start_message_id = (await state.get_data())["previous_start_message_id"]
+        await state.set_state(ManagerState.start_date)
+        await state.update_data(plane_name=message.text)
 
         current_date = datetime.now()
         current_year = current_date.year
@@ -44,7 +54,11 @@ class ManagerService:
         keyboard = await ibg.inline_buttons_generator(range(current_day, days_current_month + 1))
         keyboard.inline_keyboard.insert(0, [types.InlineKeyboardButton(text=current_month_name, callback_data=current_month_name)])
 
-        if len(keyboard.inline_keyboard) < 7:
+        count_all_inline_keyboard = 0
+        for pack_keyboard in keyboard.inline_keyboard:
+            count_all_inline_keyboard += len(pack_keyboard)
+
+        if count_all_inline_keyboard < 7:
             next_month = current_month + 1 if current_month < 12 else 1
             next_month_name = LOCAL_RU_MONTH[next_month]
             keyboard.inline_keyboard.append([types.InlineKeyboardButton(text=next_month_name, callback_data=next_month_name)])
@@ -52,11 +66,11 @@ class ManagerService:
             for day_keyboard_next_month in next_month_days.inline_keyboard:
                 keyboard.inline_keyboard.append(day_keyboard_next_month)
         else:
-            skip_row_name_math = 1
+            skip_row_name_month = 1
             max_row_buttons = 3
-            keyboard.inline_keyboard = keyboard.inline_keyboard[: max_row_buttons + skip_row_name_math]
-            keyboard.inline_keyboard[max_row_buttons + skip_row_name_math - 1].pop()
-            keyboard.inline_keyboard[max_row_buttons + skip_row_name_math - 1].pop()
+            keyboard.inline_keyboard = keyboard.inline_keyboard[: max_row_buttons + skip_row_name_month]
+            keyboard.inline_keyboard[max_row_buttons + skip_row_name_month - 1].pop()
+            keyboard.inline_keyboard[max_row_buttons + skip_row_name_month - 1].pop()
 
         temp_data_with_date_for_user = await ibg.get_string_representation_pool_inline_buttons(keyboard)
         await state.update_data(temp_data_with_date_for_user=temp_data_with_date_for_user)
@@ -64,6 +78,7 @@ class ManagerService:
             "Когда начнем? Запланируем старт ближайшую неделю. Сначала выберем подходящий день.",
             reply_markup=keyboard,
         )
+        await message.chat.delete_message(previous_start_message_id)
 
     async def survey_finish_date(self, callback_query: types.CallbackQuery, state: FSMContext):
         if callback_query.data in LOCAL_RU_MONTH.values():  # NOTE заглушка, если пользователь протыкает название месяца, а не дату.
@@ -108,9 +123,9 @@ class ManagerService:
         await state.set_state(ManagerState.finish_date)
         await callback_query.message.delete()
         await callback_query.message.answer(
-            "Теперь нужно выбрать, окончание курса приема лекарст. Заланируем на билжайший месяц. Но не более двух месяцев. "
-            "Если вы собираетесь сделать такой длинный курс, то лучше"
-            "соориентруйтесь на курс лекарств, и лучше перезапустите/повторите месячный курс по завершению первого.",
+            "Теперь нужно выбрать окончание курса.\nЗаланируем на ближайший месяц?\nНо не более двух месяцев.\n"
+            "Если же вы собираетесь сделать такой длинный курс, то лучше "
+            "соориентруйтесь на один курс, а лучше затем перезапустите/повторите месячный курс по завершению первого.",
             reply_markup=keyboard,
         )
 
@@ -143,6 +158,21 @@ class RegimenService:
         self.regimen_api_client: RegimenApiClient = regimen_api_client
         self.user_service: UserService = user_service
         self.API_URL = settings.BASE_API_URL
+        self._validate_input = validate_input
+
+    @staticmethod
+    async def __out_limit(callback_query: types.CallbackQuery, state: FSMContext, new_state_data: dict) -> bool:
+        if new_state_data["count_regimen"] > 3:
+            await callback_query.answer(
+                text=(
+                    "На один курс/плана/схему, разрешено максимум `четыре` напоминания в течении дня.\n"
+                    "Поскольку вы достигли лимита на один курс, вы можете просто создать новый.\n"
+                ),
+                show_alert=True,
+            )
+            await state.clear()
+            await callback_query.message.delete()
+            await callback_query.message.answer("Данные сохранены в системе.")
 
     async def __add_full_plane_and_first_regimen_time(self, callback_query: types.CallbackQuery, state: FSMContext):
         state_data = await state.get_data()
@@ -169,10 +199,8 @@ class RegimenService:
         await state.update_data(count_regimen=1)
         await state.update_data(manager_id=response_manager.json()["id"])
 
-    async def __add_more_regimen_time(self, callback_query: types.CallbackQuery, state: FSMContext):
+    async def __add_more_regimen_time(self, callback_query: types.CallbackQuery, state: FSMContext) -> None:
         state_data = await state.get_data()
-        if state_data["count_regimen"] >= 10:
-            raise
 
         body = {
             "manager_id": state_data["manager_id"],
@@ -181,6 +209,7 @@ class RegimenService:
             "reception_time": datetime.strptime(state_data["regimen_time"], "%H:%M").time().strftime("%H:%M"),
         }
         path = self.API_URL + "/drug-regimen/regimen/complex"
+        # logging.info(f"Запрос будет отправлен на: {path}")
         await self.manager_api_client.post_one(body, path)
         await state.update_data(count_regimen=state_data["count_regimen"] + 1)
 
@@ -197,7 +226,6 @@ class RegimenService:
 
     async def survey_regimen_hour(self, callback_query: types.CallbackQuery, state: FSMContext):
         user_choice = callback_query.data
-        # await state.update_data(regimen_hour=user_choice)
         await state.set_state(RegimenState.regimen_time)
         keyboard = await ibg.inline_buttons_generator(
             [x if x >= 10 else "0" + str(x) for x in range(0, 56, 5)],
@@ -221,6 +249,10 @@ class RegimenService:
         await state.update_data(previous_message_id=new_message.message_id)
 
     async def survey_supplement(self, message: types.Message, state: FSMContext):
+        if not self._validate_input(message.text):
+            await message.reply("Ваш ввод содержит недопустимые символы. Пожалуйста, попробуйте снова.")
+            return
+
         previous_message_id = (await state.get_data())["previous_message_id"]
         user_text = message.text
         await state.update_data(regimen_supplement=user_text)
@@ -232,15 +264,19 @@ class RegimenService:
             reply_markup=await ibg.yes_or_not_inline_buttons(),
         )
 
-    async def survey_add_more(self, callback_query: types.CallbackQuery, state: FSMContext):
+    async def survey_add_more(self, callback_query: types.CallbackQuery, state: FSMContext) -> None:
         state_data = await state.get_data()
-        # user_choice = callback_query.data
-        # await state.update_data(regimen_supplement=user_choice)
+
         if callback_query.data == "yes":
             if state_data.get("count_regimen", False):
                 await self.__add_more_regimen_time(callback_query, state)
             else:
                 await self.__add_full_plane_and_first_regimen_time(callback_query, state)
+
+            post_event_state_data = await state.get_data()
+            if await self.__out_limit(callback_query, state, post_event_state_data):
+                return None
+
             await state.set_state(RegimenState.hour)
             keyboard = await ibg.inline_buttons_generator(range(0, 24), postfix=":xx")
             await callback_query.message.delete()
@@ -253,9 +289,8 @@ class RegimenService:
                 await self.__add_more_regimen_time(callback_query, state)
             else:
                 await self.__add_full_plane_and_first_regimen_time(callback_query, state)
+
             await state.set_state(RegimenState.hour)
             await state.clear()
             await callback_query.message.delete()
-            await callback_query.message.answer(
-                "Данные сохранены в системе.",
-            )
+            await callback_query.message.answer("Данные сохранены в системе.")
